LPS-69934 Move tests to module
