+++ modules/apps/collaboration/document-library/document-library-test/src/testIntegration/java/com/liferay/document/library/subscription/test/DLSubscriptionAuthorTest.java
@@ -12,21 +12,17 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.subscriptions;
+package com.liferay.document.library.subscription.test;
 
-import com.liferay.document.library.kernel.model.DLFileEntryMetadata;
-import com.liferay.document.library.kernel.model.DLFileEntryType;
-import com.liferay.document.library.kernel.model.DLFileEntryTypeConstants;
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.document.library.kernel.service.DLAppLocalServiceUtil;
-import com.liferay.document.library.kernel.service.DLFileEntryTypeLocalServiceUtil;
-import com.liferay.dynamic.data.mapping.kernel.DDMStructure;
 import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.StringPool;
@@ -34,20 +30,18 @@ import com.liferay.portal.test.randomizerbumpers.TikaSafeRandomizerBumper;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.SynchronousMailTestRule;
 import com.liferay.portlet.documentlibrary.util.test.DLAppTestUtil;
-import com.liferay.portlet.dynamicdatamapping.util.test.DDMStructureTestUtil;
-import com.liferay.portlet.subscriptions.test.BaseSubscriptionClassTypeTestCase;
+import com.liferay.portlet.subscriptions.test.BaseSubscriptionAuthorTestCase;
 
-import org.junit.Assert;
 import org.junit.ClassRule;
 import org.junit.Rule;
+import org.junit.runner.RunWith;
 
 /**
- * @author Sergio González
- * @author Roberto Díaz
+ * @author José Ángel Jiménez
  */
+@RunWith(Arquillian.class)
 @Sync
-public class DLSubscriptionClassTypeTest
-	extends BaseSubscriptionClassTypeTestCase {
+public class DLSubscriptionAuthorTest extends BaseSubscriptionAuthorTestCase {
 
 	@ClassRule
 	@Rule
@@ -56,20 +50,18 @@ public class DLSubscriptionClassTypeTest
 			new LiferayIntegrationTestRule(), SynchronousMailTestRule.INSTANCE);
 
 	@Override
-	protected long addBaseModelWithClassType(
-			long containerModelId, long classTypeId)
+	protected long addBaseModel(long userId, long containerModelId)
 		throws Exception {
 
 		ServiceContext serviceContext =
 			ServiceContextTestUtil.getServiceContext(
-				group.getGroupId(), TestPropsValues.getUserId());
+				group.getGroupId(), userId);
 
 		DLAppTestUtil.populateNotificationsServiceContext(
 			serviceContext, Constants.ADD);
-		DLAppTestUtil.populateServiceContext(serviceContext, classTypeId);
 
 		FileEntry fileEntry = DLAppLocalServiceUtil.addFileEntry(
-			TestPropsValues.getUserId(), group.getGroupId(), containerModelId,
+			userId, group.getGroupId(), containerModelId,
 			RandomTestUtil.randomString() + ".txt", ContentTypes.TEXT_PLAIN,
 			RandomTestUtil.randomBytes(TikaSafeRandomizerBumper.INSTANCE),
 			serviceContext);
@@ -78,46 +70,27 @@ public class DLSubscriptionClassTypeTest
 	}
 
 	@Override
-	protected long addClassType() throws Exception {
-		DDMStructure ddmStructure = DDMStructureTestUtil.addStructure(
-			DLFileEntryMetadata.class.getName());
+	protected long addContainerModel(long userId, long containerModelId)
+		throws Exception {
 
 		ServiceContext serviceContext =
 			ServiceContextTestUtil.getServiceContext(
-				group.getGroupId(), TestPropsValues.getUserId());
+				group.getGroupId(), userId);
 
-		DLFileEntryType fileEntryType =
-			DLFileEntryTypeLocalServiceUtil.addFileEntryType(
-				TestPropsValues.getUserId(), group.getGroupId(),
-				RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-				new long[] {ddmStructure.getStructureId()}, serviceContext);
+		Folder folder = DLAppLocalServiceUtil.addFolder(
+			userId, group.getGroupId(), containerModelId,
+			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
+			serviceContext);
 
-		return fileEntryType.getFileEntryTypeId();
+		return folder.getFolderId();
 	}
 
 	@Override
-	protected void addSubscriptionClassType(long classTypeId) throws Exception {
-		DLAppLocalServiceUtil.subscribeFileEntryType(
-			user.getUserId(), group.getGroupId(), classTypeId);
-	}
-
-	@Override
-	protected void deleteSubscriptionClassType(long classTypeId)
+	protected void addSubscription(long userId, long containerModelId)
 		throws Exception {
 
-		DLAppLocalServiceUtil.unsubscribeFileEntryType(
-			user.getUserId(), group.getGroupId(), classTypeId);
-	}
-
-	@Override
-	protected Long getDefaultClassTypeId() throws Exception {
-		DLFileEntryType basicEntryType =
-			DLFileEntryTypeLocalServiceUtil.getDLFileEntryType(
-				DLFileEntryTypeConstants.FILE_ENTRY_TYPE_ID_BASIC_DOCUMENT);
-
-		Assert.assertNotNull(basicEntryType);
-
-		return basicEntryType.getPrimaryKey();
+		DLAppLocalServiceUtil.subscribeFolder(
+			userId, group.getGroupId(), containerModelId);
 	}
 
 	@Override
