+++ modules/apps/collaboration/message-boards/message-boards-test/src/testIntegration/java/com/liferay/message/boards/subscription/test/MBSubscriptionBaseModelTest.java
@@ -12,34 +12,40 @@
  * details.
  */
 
-package com.liferay.portlet.messageboards.subscriptions;
+package com.liferay.message.boards.subscription.test;
 
-import com.liferay.message.boards.kernel.constants.MBConstants;
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
+import com.liferay.message.boards.kernel.model.MBCategory;
 import com.liferay.message.boards.kernel.model.MBMessage;
 import com.liferay.message.boards.kernel.service.MBCategoryLocalServiceUtil;
 import com.liferay.message.boards.kernel.service.MBMessageLocalServiceUtil;
-import com.liferay.portal.kernel.portlet.PortletProvider;
-import com.liferay.portal.kernel.portlet.PortletProviderUtil;
+import com.liferay.portal.kernel.model.ResourceConstants;
+import com.liferay.portal.kernel.model.RoleConstants;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.Sync;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
+import com.liferay.portal.kernel.test.util.RoleTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.util.Constants;
+import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.SynchronousMailTestRule;
 import com.liferay.portlet.messageboards.util.test.MBTestUtil;
-import com.liferay.portlet.subscriptions.test.BaseSubscriptionLocalizedContentTestCase;
+import com.liferay.portlet.subscriptions.test.BaseSubscriptionBaseModelTestCase;
 
 import org.junit.ClassRule;
 import org.junit.Rule;
+import org.junit.runner.RunWith;
 
 /**
  * @author Roberto Díaz
  */
+@RunWith(Arquillian.class)
 @Sync
-public class MBSubscriptionLocalizedContentTest
-	extends BaseSubscriptionLocalizedContentTestCase {
+public class MBSubscriptionBaseModelTest
+	extends BaseSubscriptionBaseModelTestCase {
 
 	@ClassRule
 	@Rule
@@ -67,32 +73,42 @@ public class MBSubscriptionLocalizedContentTest
 	}
 
 	@Override
-	protected void addSubscriptionContainerModel(long containerModelId)
+	protected long addContainerModel(long userId, long containerModelId)
 		throws Exception {
 
-		MBCategoryLocalServiceUtil.subscribeCategory(
-			user.getUserId(), group.getGroupId(), containerModelId);
+		ServiceContext serviceContext =
+			ServiceContextTestUtil.getServiceContext(
+				group.getGroupId(), userId);
+
+		MBTestUtil.populateNotificationsServiceContext(
+			serviceContext, Constants.ADD);
+
+		_category = MBCategoryLocalServiceUtil.addCategory(
+			userId, containerModelId, RandomTestUtil.randomString(),
+			StringPool.BLANK, serviceContext);
+
+		return _category.getCategoryId();
 	}
 
 	@Override
-	protected String getPortletId() {
-		return PortletProviderUtil.getPortletId(
-			MBMessage.class.getName(), PortletProvider.Action.VIEW);
+	protected void addSubscriptionBaseModel(long baseModelId) throws Exception {
+		MBMessageLocalServiceUtil.subscribeMessage(
+			user.getUserId(), baseModelId);
 	}
 
 	@Override
-	protected String getServiceName() {
-		return MBConstants.SERVICE_NAME;
-	}
+	protected void removeContainerModelResourceViewPermission()
+		throws Exception {
 
-	@Override
-	protected String getSubscriptionAddedBodyPreferenceName() {
-		return "emailMessageAddedBody";
-	}
+		RoleTestUtil.removeResourcePermission(
+			RoleConstants.GUEST, MBCategory.class.getName(),
+			ResourceConstants.SCOPE_INDIVIDUAL,
+			String.valueOf(_category.getCategoryId()), ActionKeys.VIEW);
 
-	@Override
-	protected String getSubscriptionUpdatedBodyPreferenceName() {
-		return "emailMessageUpdatedBody";
+		RoleTestUtil.removeResourcePermission(
+			RoleConstants.SITE_MEMBER, MBCategory.class.getName(),
+			ResourceConstants.SCOPE_INDIVIDUAL,
+			String.valueOf(_category.getCategoryId()), ActionKeys.VIEW);
 	}
 
 	@Override
@@ -112,5 +128,7 @@ public class MBSubscriptionLocalizedContentTest
 			userId, message.getMessageId(), RandomTestUtil.randomString(),
 			serviceContext);
 	}
+
+	private MBCategory _category;
 
 }
